+ force hashing of longs to ensure consistency.
git-svn-id: https://ipaddr-py.googlecode.com/svn@196 09200d28-7f98-11dd-ad27-0f66e57d2035

+++ trunk/ipaddr.py
@@ -476,7 +476,7 @@ class _BaseIP(_IPAddrBase):
         return  '%s' % self._string_from_ip_int(self._ip)
 
     def __hash__(self):
-        return hash(hex(self._ip))
+        return hash(hex(long(self._ip)))
 
     def _get_address_key(self):
         return (self._version, self)
